♻️ [RUM-232] create a @datadog/browser-worker package
