+++ .github/codeql-config.yml
@@ -6,5 +6,6 @@ paths:
   - packages/rum/src
   - packages/rum-core/src
   - packages/rum-slim/src
+  - packages/worker/src
 paths-ignore:
   - '**/*.spec.ts'
