+++ .wokeignore
@@ -3,4 +3,4 @@ yarn.lock
 .yarn/releases
 
 # code import
-deflateWorker.js
+packages/worker/src/domain/deflate.js
