+++ packages/rum/src/domain/segmentCollection/deflateWorker.spec.ts
@@ -1,5 +1,6 @@
 import { isIE } from '@datadog/browser-core'
-import type { DeflateWorker, DeflateWorkerResponse } from './deflateWorker'
+import type { DeflateWorkerResponse } from '@datadog/browser-worker'
+import type { DeflateWorker } from './deflateWorker'
 import { createDeflateWorker } from './deflateWorker'
 
 describe('deflateWorker', () => {
