+++ packages/rum/src/domain/segmentCollection/index.ts
@@ -1,5 +1,5 @@
 export { startSegmentCollection, setSegmentBytesLimit } from './segmentCollection'
-export { DeflateWorker, DeflateWorkerAction, DeflateWorkerListener } from './deflateWorker'
+export { DeflateWorker, DeflateWorkerListener } from './deflateWorker'
 export { startDeflateWorker } from './startDeflateWorker'
 export { SEGMENT_BYTES_LIMIT } from './segmentCollection'
 export type { BrowserSegmentMetadataAndSegmentSizes } from './buildReplayPayload'
