+++ packages/rum/src/domain/segmentCollection/segment.ts
@@ -1,8 +1,9 @@
 import { addTelemetryDebug, assign, sendToExtension, addEventListener } from '@datadog/browser-core'
+import type { DeflateWorkerResponse } from '@datadog/browser-worker'
 import type { BrowserRecord, BrowserSegmentMetadata, CreationReason, SegmentContext } from '../../types'
 import { RecordType } from '../../types'
 import * as replayStats from '../replayStats'
-import type { DeflateWorker, DeflateWorkerResponse } from './deflateWorker'
+import type { DeflateWorker } from './deflateWorker'
 
 let nextId = 0
