+++ packages/worker/.npmignore
@@ -1,0 +1,4 @@
+*
+!/src/**/*
+/src/**/*.spec.ts
+/src/**/*.specHelper.ts
