+++ packages/worker/.yarnrc
@@ -1,0 +1,2 @@
+save-exact true
+
