+++ packages/worker/README.md
@@ -1,0 +1,3 @@
+# `worker`
+
+Datadog browser web Worker.
