+++ packages/worker/webpack.config.js
@@ -1,0 +1,10 @@
+const path = require('path')
+
+const webpackBase = require('../../webpack.base')
+
+module.exports = (_env, argv) =>
+  webpackBase({
+    mode: argv.mode,
+    entry: path.resolve(__dirname, 'src/entries/main.ts'),
+    filename: 'worker.js',
+  })
